tx: hoist 25 shared test helpers into tx/test_support.rs
Phase 0 of distributing tx/mod.rs::tests across the per-file sibling
modules (rings, stats, transmit, drain, cos_classify, and the cos/*
siblings). Before any test can move out of the inline mod.rs::tests
block, the 25 shared helpers (`test_queue_fast_path`,
`test_cos_runtime_with_queues`, `test_session_key`,
`build_ipv4_test_packet`, etc.) must live somewhere visible to all
sibling test modules — `mod tests { use super::*; }` only reaches
items declared above the inline block.

This commit:
- Creates `userspace-dp/src/afxdp/tx/test_support.rs` with
  `pub(in crate::afxdp)` visibility on every helper. That scope lets
  `tx/*.rs` and `cos/*.rs` siblings reach them via
  `use crate::afxdp::tx::test_support::*;` once their tests start
  moving.
- Adds `#[cfg(test)] pub(in crate::afxdp) mod test_support;` to
  `tx/mod.rs` (right after `cos_classify` so all P2 sibling decls
  cluster together).
- Replaces the 25 inline helper bodies in `tx/mod.rs::tests` with a
  single `use super::test_support::*;` line — the existing tests
  reach the helpers through the same names they always used.

Mechanical scope: 477 lines removed from mod.rs (helper bodies),
529 lines added (the new file's contents — bodies are byte-for-byte
identical to the originals modulo dedent + visibility prefix).

Build: cargo build --bins clean (one pre-existing unused-import
warning unchanged).
Tests: cargo test --bins 865/0/2 — same counts as the pre-hoist
baseline, no semantic change.

No production code changes. Helpers are still gated `#[cfg(test)]`
via the file-level `#![cfg(test)]` inner attribute and the
`#[cfg(test)]` mod declaration.
